+++ src/styles/index.ts
@@ -11,10 +11,13 @@ export { MarginCSS } from "./css"
 
 export const GlobalStyle = createGlobalStyle`
 
-body {
+body, #body {
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
     text-rendering: optimizeLegibility;
+    text-align: initial;
+    padding: initial;
+    margin: initial;
     font-size: ${FONT_SIZES["std"]};
     background-color: ${LIGHT_MODE_COLORS["background"]};
     color: ${LIGHT_MODE_COLORS["text"]};